drive get motor velocity no gets the other sides encoder position

+++ src/Drive/Drive.h
@@ -24,7 +24,7 @@ public:
 	float getTurnSpeed();
 
 	int getCANTalonEncPos();
-	float getCANTalonEncVel();
+	int getCANTalonEncVel();
 
 private:
 	CANTalon* frontLeftDrive;
